Make Cros EC skip sysjump if code runs in RAM
Sysjump isn't necessary for the ECs which execute code in RAM.

This patch makes the flashrom utility query the cros EC whether it
runs code in RAM or not. If so, the flashrom utility will assume
it's safe to erase and write the blocks even if they're part of
the regions where the EC is currently executing the code.

BUG=b:77306460
TEST=Run flashrom -p ec -w /path/to/ec.bin on Fizz.
Verify EC doesn't sysjump.

Change-Id: Iaf55d4cff63bbc4d69b91a05480d0861bd834bbc
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/994348
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
(cherry picked from commit cfd7dfc9e6c292b32e56dd2469ce73b9537b7eda)
Reviewed-on: https://chromium-review.googlesource.com/999894