Handle read-only SPI regions
If a region of the SPI flash cannot be written, then it should not be
checked against the original when it is time to verify.

The erase/write commands already handle "access denied" errors without
failing so this just extends the same support to verify.

In order to plumb this through I added a new "check_access()" function
that can be set for the flash chip.  It is only implemented for the
opaque hwseq programmer.

BUG=b:130706387
BRANCH=noe
TEST=ensure flashrom does not fail when EC region is read only

Change-Id: I4524c75105f1e2d3ee0f1bdd882c2ee1d415c0d9
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1584648
Commit-Ready: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
(cherry picked from commit 25a4ca21467783198c48474796d6c21a8fd4e08b)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1596506